+++ package.json
@@ -6,7 +6,7 @@
   "publisher": "env0",
   "main": "index.js",
   "scripts": {
-    "build": "ncc build index.js --out dist"
+    "build": "ncc build index.ts --out dist"
   },
   "keywords": [],
   "author": "",
@@ -22,7 +22,8 @@
     "husky": "^4.3.6",
     "jest": "^26.6.3",
     "nock": "^13.0.5",
-    "semistandard": "^16.0.0"
+    "semistandard": "^16.0.0",
+    "typescript": "^4.1.3"
   },
   "jest": {
     "testPathIgnorePatterns": [
